Refactor MIDI player into modular package
